requirements.txt
ncatbot>=3.8.10
jmcomic>=2.6.5
img2pdf
sqlalchemy>=2.0.43
PyYAML>=6.0.2